Add lib/swagger-scraper-test
